feat: FILES-196 - Implement upload-to API to push files in mailbox store (#4)
The upload-to API allows to upload a specific file to the carbonio
mailbox store. It returns an attachment id that can be attached to a
carbonio item (e.g. a carbonio mail draft).

The MailboxHttpClient has been implemented in order to execute the upload.

The ExceptionsHandler has been implemented to handle the exceptions in
one place. It has been attached as the last handler of the Netty
channel. For now it is used only for this API but in the future it can
be used by every APIs.