fix: FILES-263 - Fix preview/thumbnail requests in bulk
When a client makes many thumbnail requests at the same time, the system
returns mismatched images. The cause of this bug is that there is only
one shared PreviewController instance for all requests (it is a
@ChannelHandler.Sharable). This fix removes the state of the
PreviewController so there isn't the possibility that a new request will
overwrite the parameters of the previous one.

The same fix is made to the BlobController class.

Co-authored-by: RakuJa <daniele.giachetto@zextras.com>