fix: FILES-112 - Avoid creating folder with the same name (#13)
When a user tries to create a folder in a destination that already
contains a folder with the same name then the system changes the name of
the folder to create in order to avoid having duplicate folders.
Example:
 - The destination folder contains the folder: `test`
 - A user creates another folder with the same name
 - The system creates the folder calling it `test (1)`

Checked if the requester has the write permission on the destination
folder. If not, the system returns a  `NODE_WRITE_ERROR`.