fix: FILES-27 FILES-71 FILES-115 - Fix findNodes and sorting
**FILES-27**
findNodes API is now case in-sensitive: the fix is to lowercase
everything: the name and description attributes and each keyword in
input.

**FILES-71**
The sort by type is now working, so folders are always shown before
files except when the user sort by size descendent: in this case,
folders are shown after files.

**FILES-115**
When an user chose the updated_at sort, the system threw an exception
because PostgreSQL wants every attribute used in multiple joined tables
to be disambigued. For now we added an explicit 't0' before each attribute
but in the future a refactor of how the search query is created needs to
be done.