fix: CopyNodes does not maintain the name when there are conflicts (#23)
There was ambiguity in the usage of name and filename, the usage of name did not check the extension of the file, making it so that files with same name but different extensions got marked as duplicates. Another bug present was that while all the checks were done with the name, the filename was passed in the copy nodes function, and that function expected as input the name so you would have "filename.ext" that became "filename(1).ext.ext"

The fix was straight forward:
Usage of filename and not name in all the logic components of copy
The save copyNode function now expects a filename and not a name, it will never append the extension to it.