feat: FILES-290 - Implement getAccountsByEmail API (#25)
The getAccountsByEmail API allows to fetch a list of GraphQL Accounts in
batch given their email. If an email is not associated to a Carbonio
user then the system returns an Account valued at null.

Implemented the getAccountsByeEmailValidation method to validate the
emails in input and returns an error if some of them are not valid.