feat: Add user cache to reduce requests over network
It is very likely that a GraphQL response involving a Node contains
multiple User type attributes and it is very likely that these
representing the same user.

When a user is requested the first time, the system makes request to
the UserManagement over the network and caches the result. Every time
the same user is needed, the system goes directly in cache and does
not make the same network request to have the same metadata.

For now the user remains in cache for 5 minutes.