feat: Speed up fetching of children nodes
In order to speed up the fetching of the children the system now uses
the findNode functionality and returns a NodePage containing the list of
children and the PageToken necessary to request the next page (if there
are any).

The nodeCustomAttributeFetcher has been removed because the flag of each
node is retrieved during the conversion from Node to GraphQL Map. This
change allows to group the fetching of custom attributes of multiple
nodes in one single sql query.