fix: FILES-311 - Fix download timeout on large files (#31)
The service-discover has a default timeout for every requests between
services. This timeout is set to 15 seconds and it is not enough to
complete a download of files larger than around 500MB.

A new service-discover configuration has been added: service-router.json.
It will be written to the service-discover during the setup phase and it will 
override the timeout for the following endpoints:
 - /download
 - /upload
 - /upload-version

For now the timeout is increased to 1h but we saw that if, for some
reason, the download fails the service-discover close the connection
after around 60 seconds.