feat: FILES-194 - Reduce response latencies of GraphQL APIs (#30)
* feat: Add user cache to reduce requests over network

It is very likely that a GraphQL response involving a Node contains
multiple User type attributes and it is very likely that these
representing the same user.

When a user is requested the first time, the system makes request to
the UserManagement over the network and caches the result. Every time
the same user is needed, the system goes directly in cache and does
not make the same network request to have the same metadata.

For now the user remains in cache for 5 minutes.

* feat: Add GraphQL NodeBatchLoader to optimize the Nodes fetching

- Added the NodeBatchLoader in order to minimize the sql queries. This
  loader is used by the getNodeDataFetcher method to queue each node
  request in one single sql query.

- Optimized how a node is converted to a GraphQL Map<String, Object>

- Added the mutual @OneToMany, @ManyToOne annotations in the Node and
  FileVersion entities. This will help ebean to group same queries
  in a single one when is possible.

* feat: Add GraphQL ShareBatchLoader to optimize the Shares fetching

- Added the ShareBatchLoader in order to minimize the sql queries. This
  loader is used by the getSharesDataFetcher method to queue a list
  of shares to retrieve in one single sql query.

- Added a getShares(List<String> nodeIds) method in ShareRepository.
  It is necessary to fetch all the shares for each node ids.

* feat: Replace custom cache with Ebean cache

- Replaced the Node and Share custom caches with the Ebean cache using
  the @Cache annotation. Every logic about these two caches has been
  removed from their respective Repository.

- Fixed the FileVersion cache (now it is much more used and cleaned).
  It remains until we figure out how to replace it with Ebean.

- Removed the @Transactional annotations where not necessary.

For now, the Ebean cache has the following limits:
- MaxSize: 100_000 items
- MaxTimeToLive: 300 seconds
- MaxIdleTime: 300 seconds

These are very conservative settings but it should be neccesary to speed
up close requests that require the same items.

* feat: Speed up fetching of children nodes

In order to speed up the fetching of the children the system now uses
the findNode functionality and returns a NodePage containing the list of
children and the PageToken necessary to request the next page (if there
are any).

The nodeCustomAttributeFetcher has been removed because the flag of each
node is retrieved during the conversion from Node to GraphQL Map. This
change allows to group the fetching of custom attributes of multiple
nodes in one single sql query.