fix: FILES-354 - Fix GetChildren API called on LOCAL_ROOT folder (#32)
Now the GetChildren API called on the LOCAL_ROOT folder returns
only the nodes owned by the requester. For all other folders it
continues to return all nodes which the requester has the permission:
the owned nodes and the sharedWith him/her nodes.

Fix also the conversion from Node to GraphQL map: now it converts the
right version of the Node (when it is a File and it is necessary to
convert a FileVersion that is not the last one)