fix: FILES-437 - Fix timeout of copy API when file to copy is large
When a user tries to copy a file with a large size, the copy API goes in timeout.
Now the system uses the Filestore' copy method instead of downloading and uploading
the blob manually. This is a temporary workaround because this problem can be
resolved only making the API asynchronous.

When the copy fails now the system handles the failure returning a specific
GraphQLError: NODE_COPY_ERROR.