feat: FILES-484 - Add cache-control and etag headers in PreviewController
In PreviewController the following header have been added on each HTTP response:
 - cache-control: no-cache
 - etag: <blob-digest>

When a client specifies the IF_NONE_MATCH header containing a digest (our etag value)
the system returns a 304 status code if the the requested blob to preview has the same digest.
Otherwise it returs the requested preview with the etag header containing its digest.

These headers allow to handle the client caching mechanism:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/ETag)