fix: FILES-618 Add ExceptionHandler at the end of each endpoint pipeline
The ExceptionHandler was introduced months ago but it was applied only to some
endpoints. This fix added the handler at the end of each endpoint pipeline.
A micro refactor was done to redirect all the exception to
this handler but I consider it as a starting point for a major refactor
that needs to be done in the future