feat: make node version as a query param in Preview APIs (#151)
This technical improvement allows the client to call the preview and
thumbnail APIs without being forced to specify the file version. It is
useful when the client want the preview of the latest file version
without knowing which is. Now the version of the file is optional and
can be passed as a query param `?version=1` or `&version=1`

- Added PreviewIT to cover all the success cases
- Added ThumbnailIT to cover all the success cases
- Improved the PreviewController to automatically fetch the latest
  version of the file to preview if the node version is not passed

Refs: FILES-227